requirements.txt
hugchat==0.0.6.2
pysftp==0.2.9
python-dotenv==1.0.0
python-telegram-bot==20.3
